Move reCAPTCHA assets to a new file
